requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
joblib==1.3.2
pydantic==2.6.3
mlflow==2.13.0
setuptools>=70.0.0
pytest==7.4.3
httpx==0.25.0
torch
torchvision
Pillow
python-multipart